[3단계 - 데이터베이스 적용] 부산대 BE 김성문 미션 제출합니다. (#281)
* docs: 구현할 기능 목록 정리

* refactor: 데이터베이스 마이그레이션

- HashMap -> H2 데이터베이스
- 정팩메 컨벤션에 따라 `.of()`, `.from()` 구분
- 객체 ID 부여 로직을 서비스가 담당하도록 수정
- 동시성 문제 해결을 위해 `AtomicLong` 도입
- 생성, 수정을 동시에 핸들링하기 위해 MERGE문 적용
- ReadProductResponse -> ProductResponse
- 상품 생성, 수정에 대한 예외 정의 및 핸들링

* feat: SQL 스크립트 작성

* chore: 객체의 id 부여를 데이터베이스가 책임지도록 수정

* chore: 데이터베이스를 명시적으로 초기화하도록 수정

* refactor: id가 null인 객체를 생성할 수 있도록 정팩메 오버로딩

* refactor: 리포지토리 리팩토링

- 쿼리 가독성 향상을 위해 raw string 적용
- 엔티티 객체의 id를 데이터베이스가 부여하도록 쿼리 수정
- save 메서드가 데이터베이스에서 부여한 엔티티 id를 리턴하도록 수정
- update 메서드 구현
- 예외 처리를 위해 delete 메서드가 query affected row를 리턴하도록 수정

* refactor: 서비스 리팩토링

- 엔티티 객체 ID 부여 책임을 데이터베이스가 갖도록 수정
- `수정 및 삭제` 메서드가 semantic한 예외를 발생시킬 수 있도록 수정
- `@Transactional` 어노테이션 추가

* refactor: 관리자 페이지가 올바른 HTTP 메서드를 사용하도록 수정

+++ src/main/java/gift/service/ProductService.java
@@ -3,13 +3,18 @@ package gift.service;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import gift.domain.Product;
 import gift.dto.CreateProductRequest;
 import gift.dto.CreateProductResponse;
-import gift.dto.ReadProductResponse;
+import gift.dto.ProductResponse;
 import gift.dto.UpdateProductRequest;
 import gift.dto.UpdateProductResponse;
+import gift.exception.ProductCreateException;
+import gift.exception.ProductDeleteException;
 import gift.exception.ProductNotFoundException;
+import gift.exception.ProductUpdateException;
 import gift.repository.ProductRepository;
 
 @Service
@@ -21,30 +26,57 @@ public class ProductService {
         this.productRepository = productRepository;
     }
 
-    public List<ReadProductResponse> getAllProducts() {
+    @Transactional(readOnly = true)
+    public List<ProductResponse> getAllProducts() {
         return productRepository.findAll()
             .stream()
-            .map(ReadProductResponse::of)
+            .map(ProductResponse::from)
             .toList();
     }
 
-    public ReadProductResponse getProductById(Long id) {
-        return ReadProductResponse.of(
+    @Transactional(readOnly = true)
+    public ProductResponse getProductById(Long id) {
+        return ProductResponse.from(
             productRepository.findById(id)
                 .orElseThrow(() -> new ProductNotFoundException("해당 상품이 존재하지 않습니다.")));
     }
 
+    @Transactional
     public CreateProductResponse createProduct(CreateProductRequest request) {
-        return CreateProductResponse.of(
-            productRepository.save(request.name(), request.price(), request.imageUrl()));
+        Product product = Product.of(request.name(), request.price(), request.imageUrl());
+
+        Long generatedId = productRepository.save(product);
+
+        return CreateProductResponse.from(
+            productRepository.findById(generatedId)
+                .orElseThrow(() -> new ProductCreateException("상품 생성을 실패했습니다.")));
     }
 
+    @Transactional
     public UpdateProductResponse updateProduct(Long id, UpdateProductRequest request) {
-        return UpdateProductResponse.of(
-            productRepository.update(id, request.name(), request.price(), request.imageUrl()));
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("해당 상품이 존재하지 않습니다.");
+        }
+
+        Product newProduct = Product.of(id, request.name(), request.price(), request.imageUrl());
+
+        int count = productRepository.update(newProduct);
+        if (count != 1) {
+            throw new ProductUpdateException("상품 수정을 실패했습니다.");
+        }
+
+        return UpdateProductResponse.from(newProduct);
     }
 
+    @Transactional
     public void deleteProduct(Long id) {
-        productRepository.delete(id);
+        if (!productRepository.existsById(id)) {
+            throw new ProductNotFoundException("해당 상품이 존재하지 않습니다.");
+        }
+
+        int count = productRepository.delete(id);
+        if (count != 1) {
+            throw new ProductDeleteException("상품 삭제를 실패했습니다.");
+        }
     }
 }
